backend/requirements.txt
# =============================================================================
# PageMade Backend Dependencies
# =============================================================================
#
# Installation:
#   Production:  pip install -r requirements.txt
#   Development: pip install -r requirements.txt && pip install -r requirements.txt[dev]
#   Testing:     pip install -r requirements.txt && pip install -r requirements.txt[test]
#
# Or install all at once:
#   pip install Flask==2.3.3 Flask-SQLAlchemy==3.0.5 ... (all below)
#
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-cors==6.0.1
python-dotenv==1.0.0

# -----------------------------------------------------------------------------
# Authentication & Security
# -----------------------------------------------------------------------------
Authlib==1.2.1
requests==2.31.0
PyJWT==2.8.0
Flask-JWT-Extended==4.5.3

# -----------------------------------------------------------------------------
# Database & Storage
# -----------------------------------------------------------------------------
psycopg2-binary==2.9.7          # PostgreSQL adapter (production)
# Note: SQLite is built-in, no package needed for local dev

# -----------------------------------------------------------------------------
# Cache & Performance
# -----------------------------------------------------------------------------
redis==5.0.0

# -----------------------------------------------------------------------------
# Media Processing
# -----------------------------------------------------------------------------
Pillow==10.0.1

# -----------------------------------------------------------------------------
# Production Server
# -----------------------------------------------------------------------------
gunicorn==21.2.0

# =============================================================================
# DEVELOPMENT DEPENDENCIES (Optional)
# =============================================================================
# Uncomment these for local development:
#
# Code Quality:
# black==23.11.0
# flake8==6.1.0
# pylint==3.0.2
# mypy==1.7.1
#
# Debugging:
# flask-debugtoolbar==0.13.1
# ipython==8.18.1

# =============================================================================
# TESTING DEPENDENCIES (Optional)
# =============================================================================
# For running tests, install these:
#
# pytest==7.4.3
# pytest-cov==4.1.0
# pytest-flask==1.3.0
# pytest-mock==3.12.0
# coverage==7.3.2
# faker==20.1.0
#
# Or create a separate requirements-dev.txt with all dev+test deps


